Add files via upload
This is example code I made that generates random insults all in main method. For our project we would have to separate it into different methods of course and add whatever else we need. We could also choose to make it a compliment and insult generator if you guys want to add more to our program.